docs(vitepress): word selection feature
